plugins/data-reconciliation/requirements.txt
# Core data processing
pandas>=2.0.0
numpy>=1.24.0

# Database connectivity
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0  # PostgreSQL
pymysql>=1.0.0          # MySQL
pyodbc>=4.0.0           # SQL Server

# File formats
openpyxl>=3.1.0         # Excel read/write
xlsxwriter>=3.1.0       # Excel formatting
pyarrow>=12.0.0         # Parquet files

# API and web
requests>=2.31.0

# AI integration
google-generativeai>=0.3.0

# CLI and utilities
click>=8.1.0
tqdm>=4.65.0            # Progress bars
colorama>=0.4.6         # Colored terminal output
tabulate>=0.9.0         # Pretty tables

# Data quality and fuzzy matching
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Logging
python-json-logger>=2.0.0

# Testing (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
